# contracts/openapi/ESG_P0_OpenAPI_v0.1.yaml
openapi: 3.0.3
info:
  title: ESG报告软件 P0 API
  version: 0.1.0
  description: ESG报告软件MVP阶段P0接口规格，覆盖多租户、项目、同行报告解析、议题推荐、采集审核、知识库、AI写作校对、导出、AI成本与审计。
servers:
- url: /api/v1
  description: 默认API前缀
...
paths:
...
  /enterprises:
    post:
      tags:
      - 企业管理
      summary: 创建企业
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Enterprise'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnterpriseCreateRequest'
      security:
      - bearerAuth: []
    get:
      tags:
      - 企业管理
      summary: 查询企业列表
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnterpriseListResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      parameters:
      - name: keyword
        in: query
        required: false
        schema:
          type: string
      - name: exchange
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
      security:
      - bearerAuth: []
  /enterprises/{enterprise_id}:
    parameters:
    - name: enterprise_id
      in: path
      required: true
      schema:
        type: string
    get:
      tags:
      - 企业管理
      summary: 获取企业详情
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Enterprise'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
    patch:
      tags:
      - 企业管理
      summary: 更新企业
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Enterprise'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnterpriseCreateRequest'
      security:
      - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required:
      - success
      - message
      - error
      properties:
        success:
          type: boolean
          example: false
        data:
          nullable: true
        message:
          type: string
        error:
          $ref: '#/components/schemas/ErrorObject'
        request_id:
          type: string
        timestamp:
          type: string
          format: date-time
    ErrorObject:
      type: object
      properties:
        code:
          type: string
        type:
          type: string
        details:
          type: object
          additionalProperties: true
    PagedResponse:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
...
    EnterpriseCreateRequest:
      type: object
      required:
      - enterprise_name
      properties:
        enterprise_name:
          type: string
        enterprise_short_name:
          type: string
        stock_code:
          type: string
        exchange:
          type: string
        country_or_region:
          type: string
        industry_description:
          type: string
        main_business:
          type: string
        status:
          type: string
          enum:
          - active
          - inactive
    Enterprise:
      type: object
      properties:
        enterprise_id:
          type: string
        tenant_id:
          type: string
        enterprise_name:
          type: string
        enterprise_short_name:
          type: string
        stock_code:
          type: string
        exchange:
          type: string
        country_or_region:
          type: string
        industry_description:
          type: string
        main_business:
          type: string
        gics:
          $ref: '#/components/schemas/GicsClassification'
        status:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    EnterpriseListResponse:
      allOf:
      - $ref: '#/components/schemas/PagedResponse'
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Enterprise'
    GicsClassification:
      type: object
      properties:
        gics_level_1_code:
          type: string
        gics_level_1_name:
          type: string
        gics_level_2_code:
          type: string
        gics_level_2_name:
          type: string
        gics_level_3_code:
          type: string
        gics_level_3_name:
          type: string
        gics_level_4_code:
          type: string
        gics_level_4_name:
          type: string
...